wibestore_backend/pyproject.toml
[project]
name = "wibestore-backend"
version = "1.0.0"
description = "WibeStore - Marketplace for Gaming Accounts"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "WibeStore Team" }]

dependencies = [
    "Django>=5.1,<6.0",
    "djangorestframework>=3.15,<4.0",
    "djangorestframework-simplejwt>=5.3,<6.0",
    "django-cors-headers>=4.3,<5.0",
    "django-filter>=24.0",
    "drf-spectacular>=0.27,<1.0",
    "psycopg[binary]>=3.1,<4.0",
    "redis>=5.0,<6.0",
    "celery>=5.3,<6.0",
    "django-celery-beat>=2.6,<3.0",
    "channels>=4.2,<5.0",
    "channels-redis>=4.2,<5.0",
    "Pillow>=10.0,<11.0",
    "django-storages>=1.14,<2.0",
    "boto3>=1.34,<2.0",
    "gunicorn>=22.0,<23.0",
    "uvicorn[standard]>=0.29,<1.0",
    "django-environ>=0.11,<1.0",
    "cryptography>=42.0,<43.0",
    "django-axes>=6.4,<7.0",
    "argon2-cffi>=23.1,<24.0",
    "social-auth-app-django>=5.4,<6.0",
    "sentry-sdk[django]>=1.40,<2.0",
    "django-health-check>=3.18,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-cov>=4.1,<5.0",
    "pytest-asyncio>=0.23,<1.0",
    "factory-boy>=3.3,<4.0",
    "faker>=22.0,<23.0",
    "coverage>=7.4,<8.0",
    "black>=24.0,<25.0",
    "ruff>=0.2,<1.0",
    "isort>=5.13,<6.0",
    "pre-commit>=3.6,<4.0",
    "django-debug-toolbar>=4.3,<5.0",
    "bandit>=1.7,<2.0",
    "ipython>=8.20,<9.0",
]

[tool.black]
line-length = 99
target-version = ["py312"]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "SIM"]

[tool.isort]
profile = "black"
line_length = 99

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.testing"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
]

[tool.coverage.run]
source = ["apps", "core"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
